fix(Entity) : repository와 Entity클래스 테스트 분리
